haversine formular added. to be mordified later
added room for haversine formular to determine location and change meter read widgets accordingly, fixed the issue of account reloading, i.e, it is caused by the kivy dictionary that sometimes fail to clear. especiialy after  logging in another uiser.
to be removed later are: the textbox added to test the haversine,